Add files via upload
My solution was mostly similar to Mosh's.  Except that he did the filtering before the pagination function was called.  In my solution the filtering was done in the item click event handler so the value passed to paginate never changed.   Unsure why I got all the movies or none though.